backend/requirements.txt
amqp==5.0.2
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudinary==1.24.0
dj-database-url==0.5.0
Django==3.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-on-heroku==1.1.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
joblib==0.17.0
kombu==5.0.2
numpy==1.19.4
Pillow==8.0.1
profanity-check==1.0.3
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
redis==3.5.3
requests==2.25.1
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
